+++ CHANGELOG.md
@@ -5,6 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- The property 'Expression' in 'AADSyncRules.AttributeFlowMappings' cannot be $null.
+
 ## [0.1.0] - 2024-07-04
 
 ### Changed